+++ lib/pages/homeuser_screen.dart
@@ -18,11 +18,10 @@ class _HomeUserScreenState extends State<HomeUserScreen>
   List data = [];
   var remarks;
   int radiovalue;
-  String nama = 'Full Name';
-  String nopeg = '582813';
-  String unit = 'JTKTOB-3';
-  String email = 'email@gmf-aeroasia.co.id';
-  String voucher_amount = '4000';
+  String nama = ' ';
+  String nopeg = ' ';
+  String unit = ' ';
+  String email = ' ';
   /////////////////////////////////////////////////////////////////////////////////////////////////
   ///Controller
   /////////////////////////////////////////////////////////////////////////////////////////////////
@@ -163,9 +162,33 @@ class _HomeUserScreenState extends State<HomeUserScreen>
   }
 
   Future banteklogout() async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    prefs.clear();
-    Bantek.goToLogin(context);
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        // return object of type Dialog
+        return AlertDialog(
+          title: new Text("Logout User"),
+          content: new Text("Apa kamu yakin mau logout?"),
+          actions: <Widget>[
+            // usually buttons at the bottom of the dialog
+            new FlatButton(
+              child: new Text("Tidak"),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+            new FlatButton(
+              child: new Text("Yes"),
+              onPressed: () async {
+                SharedPreferences prefs = await SharedPreferences.getInstance();
+               prefs.clear();
+                Bantek.goToLogout(context);
+              },
+            ),
+          ],
+        );
+      },
+    );
   }
 
   void bantekform() {
@@ -177,10 +200,40 @@ class _HomeUserScreenState extends State<HomeUserScreen>
   /////////////////////////////////////////////////////////////////////////////////////////////////
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return WillPopScope(
+      child: Scaffold(
         body: Column(
       children: <Widget>[profiling(),horizontalline(), buttongroup(), listbantek()],
-    ));
+    )), onWillPop: () {
+      showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        // return object of type Dialog
+        return AlertDialog(
+          title: new Text("Logout User"),
+          content: new Text("Apa kamu yakin mau logout?"),
+          actions: <Widget>[
+            // usually buttons at the bottom of the dialog
+            new FlatButton(
+              child: new Text("Tidak"),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+            new FlatButton(
+              child: new Text("Yes"),
+              onPressed: () async {
+                SharedPreferences prefs = await SharedPreferences.getInstance();
+               prefs.clear();
+                Bantek.goToLogout(context);
+              },
+            ),
+          ],
+        );
+      },
+    );
+    },
+    );
   }
 
   Widget profiling() {
@@ -374,7 +427,10 @@ class _HomeUserScreenState extends State<HomeUserScreen>
 
                   subtitle: Column(
                     children: <Widget>[
-                      Row(
+                      SizedBox(
+                        width: 350,
+                        child: Wrap(
+                        direction: Axis.horizontal,
                         children: <Widget>[
                           Text(
                             data[index]['departure_station'],
@@ -382,38 +438,44 @@ class _HomeUserScreenState extends State<HomeUserScreen>
                                 fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),
                           data[index]['leg_st_1']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_st_1'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),   
                           data[index]['leg_st_2']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_st_2'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),  
                           data[index]['leg_st_3']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_st_3'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),    
                           data[index]['leg_st_4']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_st_4'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),  
                           data[index]['leg_st_5']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_st_5'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),                 
                         ],
                       ),
-                      Row(
+                      ),
+                      SizedBox(
+                        height: 10,
+                      ),
+                      SizedBox(
+                        width: 350,
+                        child: Wrap(
                         children: <Widget>[
                           Text(
                             data[index]['departure_city'],
@@ -421,36 +483,37 @@ class _HomeUserScreenState extends State<HomeUserScreen>
                                 fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),
                           data[index]['leg_city_1']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_city_1'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),   
                           data[index]['leg_city_2']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_city_2'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),
                           data[index]['leg_city_3']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_city_3'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0), 
                           data[index]['leg_city_4']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_city_4'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),  
                           data[index]['leg_city_5']!='null'?
-                          Row(
+                          Wrap(
                             children: <Widget>[
                               Image.asset('assets/iconairtport.png',width: 32.0,height: 12.0,),
                           Text(data[index]['leg_city_5'],style: TextStyle(fontStyle: FontStyle.italic, fontSize: 15.0),
                           ),],):Container(width: 0, height: 0),                                
                         ],
+                      ),
                       ),
                       Row(
                         children: <Widget>[
@@ -589,14 +652,14 @@ class _HomeUserScreenState extends State<HomeUserScreen>
 
   Color colorstatus(sppd, ticket, invoice, voucher, aml, status) {
     if (status != 'Y') {
-      return Colors.red;
+      return Colors.red[200];
     } else {
       if (sppd != '0' &&
           ticket != '0' &&
           invoice != '0' &&
           voucher != '0' &&
           aml != '0') {
-        return Colors.green;
+        return Colors.green[200];
       } else {
         return Colors.yellow;
       }
